+++ src/schemas/constants.py
@@ -11,7 +11,7 @@ ARRAY_OF_STRINGS = {
 
 FIELD_STRING_TYPE = {
     "type": "string",
-    "pattern": "^([^\\.]+)*?([^\\.\\d]+(\\d+)\\.{2,3}(\\d+))*?$",
+    "pattern": "^([^\\.]+|[^\\.\\d]+\\d+\\.{2,3}\\d+)$",
 }
 
 FIELD_STRING_REGEX_GROUPS = r"([^\.\d]+)(\d+)\.{2,3}(\d+)"
