+++ src/utils/parsing.py
@@ -1,5 +1,6 @@
 import re
 from copy import deepcopy
+from fractions import Fraction
 
 from deepmerge import Merger
 from dotmap import DotMap
@@ -46,6 +47,14 @@ def parse_field_string(field_string):
         ]
     else:
         return [field_string]
+
+
+def parse_float_or_fraction(result):
+    if type(result) == str and "/" in result:
+        result = float(Fraction(result))
+    else:
+        result = float(result)
+    return result
 
 
 def parse_fields(key, fields):
